Adding logic to userControl
Implemented the log-in and sign-up methods
@PersistenceContext is used to inject the EntityManager, which is responsible for interacting with the database. The EntityManager is used to create a query that selects the User entity where the name and password match the given parameters. The getSingleResult() method is used to execute the query and retrieve the result, which should be a single User object if the login is successful.

The SignUp function checks the value of the role parameter. If it is equal to "runner" (case-insensitive), it prompts the user for the delivery fees and creates a new Runner object. Otherwise, it creates a new User object. The properties are set accordingly, and the objects are persisted using the EntityManager's persist method.

in PromptForDeliveryFees() the method uses Scanner to read user input from the console. It prompts the user with the message "Please set the minimum fees for the next trip: " and reads a float value using scanner.nextFloat(). After capturing the float value, scanner.nextLine() is used to consume the newline character left in the input buffer.

Added a field called email in class user including the setters and getters. We will use this as a unique not null field to specify the users.