Updated the Log in in userControl
fixing the code due to some Error messages.

added some code to operApp method.

I added a try-catch block around the code that may throw the IllegalArgumentException. If the exception is thrown, it will be caught by the catch block. Inside the catch block, you can define the actions to be taken when the exception occurs.