Converting Runner Controller to RESTful Api
the acceptOrder and cancelOrder methods are annotated with @PUT to handle HTTP PUT requests. The @Path annotations on the methods specify the additional path segments for the respective endpoints. The @PathParam annotations extract the orderId from the URL path as a method parameter.

 the completeOrder method is annotated with @PUT and the getTripsCount method is annotated with @GET. The @Path annotations on the methods specify the additional path segments for the respective endpoints. The @PathParam annotation extracts the orderId from the URL path as a method parameter in the completeOrder method.

the getTripsHistory method is annotated with @GET and @Path("/history"). The @Path annotation specifies the additional path segment for the endpoint.

Created the CostumerInterface to implement the GUI later. And fixed the bug in UserControl.